+++ Source/Config/AppUrls.h
@@ -1,0 +1,12 @@
+#pragma once
+
+namespace svc::urls
+{
+
+inline constexpr const char* kGitHubRepo     = "https://github.com/azazhang/Super_VelocityCurve";
+inline constexpr const char* kHomepage       = "https://azhang.eu.org/";
+inline constexpr const char* kYouTube        = "https://www.youtube.com/@JJJB";
+inline constexpr const char* kSupport        = "https://buymeacoffee.com/azazhang";
+inline constexpr const char* kPluginvalInfo  = "https://www.tracktion.com/develop/pluginval";
+
+} // namespace svc::urls